+ DRY auth-related and other AppModel code. + Make admin app auth work like all other apps. + Temporarirly remove non-working non-SSO model
