+++ client-app/src/examples/todo/AppModel.ts
@@ -1,25 +1,14 @@
-import {HoistAppModel, XH} from '@xh/hoist/core';
-import {OauthService} from '../../core/svc/OauthService';
+import {XH} from '@xh/hoist/core';
 import {TaskService} from './TaskService';
+import {BaseAppModel} from '../../BaseAppModel';
 
 export const PERSIST_APP = {localStorageKey: 'todoAppState'};
 
-export class AppModel extends HoistAppModel {
+export class AppModel extends BaseAppModel {
     static instance: AppModel;
 
-    static override async preAuthAsync() {
-        await XH.installServicesAsync(OauthService);
-    }
-
     override async initAsync() {
+        await super.initAsync();
         await XH.installServicesAsync(TaskService);
-    }
-
-    override async logoutAsync() {
-        await XH.oauthService.logoutAsync();
-    }
-
-    override get supportsVersionBar(): boolean {
-        return window.self === window.top;
     }
 }
